requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
flask-sqlalchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
boto3==1.13.7
pandas==0.25.3
numpy==1.18.5
PyMySQL==0.9.3
botocore==1.16.20
seaborn==0.10.1
matplotlib==3.2.1
cycler==0.10.0
datetime==4.3
scikit-learn==0.21.3
joblib==0.15.1
pytest==5.4.1